requirements_full.txt
streamlit>=1.32.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
scikit-learn>=1.4.0
pandas>=2.0.0
numpy>=1.26.0
scipy>=1.11.0
matplotlib>=3.7.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-whois>=0.9.4
tldextract>=5.1.0
